requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aioitertools==0.12.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
boto3==1.34.131
botocore==1.34.131
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
deeplake==3.9.25
dill==0.3.9
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
humbug==0.3.2
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-text-splitters==0.3.0
langcodes==3.4.1
langsmith==0.1.132
language_data==1.2.0
load-dotenv==0.1.0
lz4==4.3.3
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.0
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.9.1
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.51.1
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathos==0.3.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
sentence-transformers==3.1.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.3
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
